Added: Requirements.txt. Updated: Cambio de nombre en los archivos:  log_in -> inicia_sesion, register -> Registrate
